fix: component system and clean reuse
